Fix corrected entries and update thermo builder
