+++ emmet-core/emmet/core/thermo.py
@@ -1,11 +1,11 @@
 """ Core definition of a Thermo Document """
 from collections import defaultdict
-from typing import Dict, List, Union
+from typing import Dict, List, Optional, Union
 from datetime import datetime
 from emmet.core.utils import ValueEnum
 
 from pydantic import BaseModel, Field
-from pymatgen.analysis.phase_diagram import PhaseDiagram, PatchedPhaseDiagram
+from pymatgen.analysis.phase_diagram import PhaseDiagram
 from pymatgen.entries.computed_entries import ComputedEntry, ComputedStructureEntry
 
 from emmet.core.material_property import PropertyDoc
@@ -106,26 +106,11 @@ class ThermoDoc(PropertyDoc):
         cls,
         entries: List[Union[ComputedEntry, ComputedStructureEntry]],
         thermo_type: Union[ThermoType, RunType],
+        phase_diagram: Optional[PhaseDiagram] = None,
         **kwargs
     ):
-        # Note that PatchedPhaseDiagram construct the hull using only the
-        # lowest energy entries.
-        patched_pd = PatchedPhaseDiagram(entries, keep_all_spaces=True)
 
-        # Main PD of parent chemsys
-        try:
-            pd = patched_pd.pds[frozenset(patched_pd.elements)]
-        except KeyError:
-            entries_by_comp = defaultdict(list)
-            for e in entries:
-                entries_by_comp[e.composition.reduced_formula].append(e)
-
-            # Only use lowest entry per composition to speed up QHull in Phase Diagram
-            reduced_entries = [
-                sorted(comp_entries, key=lambda e: e.energy_per_atom)[0] for comp_entries in entries_by_comp.values()
-            ]
-            pd = PhaseDiagram(reduced_entries)
-            patched_pd.pds.update({frozenset(patched_pd.elements): pd})
+        pd = phase_diagram or cls.construct_phase_diagram(entries)
 
         docs = []
 
@@ -213,20 +198,35 @@ class ThermoDoc(PropertyDoc):
 
             docs.append(ThermoDoc.from_structure(meta_structure=blessed_entry.structure, **d, **kwargs))
 
-        # Construct new phase diagrams with all of the entries, not just those on the hull
-        new_pds = []
-        for ele_set, pd in patched_pd.pds.items():
-            new_entries = []
-            for entry in entries:
-                if frozenset(entry.composition.elements).issubset(ele_set):
-                    new_entries.append(entry)
+        return docs
 
-            pd_computed_data = pd.computed_data
-            pd_computed_data["all_entries"] = new_entries
-            new_pd = PhaseDiagram(new_entries, elements=pd.elements, computed_data=pd_computed_data)
-            new_pds.append(new_pd)
+    @staticmethod
+    def construct_phase_diagram(entries) -> PhaseDiagram:
+        """
+        Efficienty construct a phase diagram using only the lowest entries at every composition
+        represented in the entry data passed.
 
-        return docs, new_pds
+        Args:
+            entries (List[ComputedStructureEntry]): List of corrected pymatgen entry objects.
+
+        Returns:
+            PhaseDiagram: Pymatgen PhaseDiagram object
+        """
+        entries_by_comp = defaultdict(list)
+        for e in entries:
+            entries_by_comp[e.composition.reduced_formula].append(e)
+
+        # Only use lowest entry per composition to speed up QHull in Phase Diagram
+        reduced_entries = [
+            sorted(comp_entries, key=lambda e: e.energy_per_atom)[0] for comp_entries in entries_by_comp.values()
+        ]
+        pd = PhaseDiagram(reduced_entries)
+
+        # Add back all entries, not just those on the hull
+        pd_computed_data = pd.computed_data
+        pd_computed_data["all_entries"] = entries
+        new_pd = PhaseDiagram(entries, elements=pd.elements, computed_data=pd_computed_data)
+        return new_pd
 
 
 class PhaseDiagramDoc(BaseModel):
